# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/execute": {
      "post": {
        "operationId": "ExecuteController_execute",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Ejecución completada"
          },
          "403": {
            "description": "Tool fuera de la whitelist del Executor"
          },
          "502": {
            "description": "Fallo real de la API de Modal (tier remoto)"
          },
          "504": {
            "description": "El pod no terminó dentro del timeout (tier local)"
          }
        },
        "summary": "Ejecuta código en un pod Deno aislado (language: typescript) o en un sandbox de Modal (language: python) — la decisión de tier es automática según el language, no la declara el caller (BLUEPRINT 4.5)",
        "tags": [
          "execute"
        ]
      }
    },
    "/services": {
      "post": {
        "operationId": "PreviewServiceController_start",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Servicio creado"
          },
          "403": {
            "description": "Tool fuera de la whitelist del Executor"
          },
          "429": {
            "description": "Límite de servicios concurrentes alcanzado"
          }
        },
        "summary": "Levanta un pod de servicio de larga vida expuesto bajo https://<slug>.jinserver.com (Fase 5.5)",
        "tags": [
          "services"
        ]
      },
      "get": {
        "operationId": "PreviewServiceController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Lista los pods de servicio activos y su TTL restante",
        "tags": [
          "services"
        ]
      }
    },
    "/services/{id}": {
      "delete": {
        "operationId": "PreviewServiceController_stop",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Detiene y destruye un pod de servicio activo",
        "tags": [
          "services"
        ]
      }
    },
    "/terminal/sessions": {
      "post": {
        "operationId": "TerminalController_start",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Abre una sesión de terminal aislada (ADR 0016)",
        "tags": [
          "terminal"
        ]
      },
      "get": {
        "operationId": "TerminalController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Sesiones de terminal activas",
        "tags": [
          "terminal"
        ]
      }
    },
    "/terminal/sessions/{id}": {
      "delete": {
        "operationId": "TerminalController_stop",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Cierra una sesión y destruye su pod",
        "tags": [
          "terminal"
        ]
      }
    },
    "/terminal/sessions/{id}/exec": {
      "post": {
        "operationId": "TerminalController_exec",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Ejecuta un comando en la sesión (salida en streaming NDJSON)",
        "tags": [
          "terminal"
        ]
      }
    },
    "/terminal/sessions/{id}/files": {
      "get": {
        "operationId": "TerminalController_exportFiles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Archivos de texto del espacio de trabajo (para traerlos al editor)",
        "tags": [
          "terminal"
        ]
      },
      "put": {
        "operationId": "TerminalController_importFiles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Copia archivos del editor al espacio de trabajo",
        "tags": [
          "terminal"
        ]
      }
    },
    "/terminal/sessions/{id}/expose": {
      "post": {
        "operationId": "TerminalController_expose",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Publica un directorio de la sesión bajo https://<slug>.jinserver.com",
        "tags": [
          "terminal"
        ]
      }
    }
  },
  "info": {
    "title": "Jin Executor API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {}
  }
}